requirements.txt
gpiozero>=1.6     # 高階 GPIO 控制（取代 RPi.GPIO，支援 Pi 5）
lgpio             # Pi 5 使用 gpiozero 時所需的底層驅動（自動載入）

PyYAML>=6.0       # 讀取 config.yaml
python-tm1637>=1.1.1
python-kasa
